requirements.txt
requests
ijson
Pillow